+++ npm/metrics/endpoints_windows.go
